fix: move scene ui fields to update cycle (#5)
Resolves bug discussed in https://github.com/excaliburjs/Excalibur/discussions/2399

Line 705 of dev-tools is slightly heavy, but it needs to account for both dynamic scenes added and dynamic scenes added _and_ activated. I've demonstrated this in the last example in `index.ts`. Even with a lot of scenes, I think this check would still be quick enough for development. 

In the examples, pressing "N" will add a scene, but not activate it, and pressing "A" will both add and activate a scene. In both cases we expect to see the scene list in the dev tools ui to update.